pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videoteka-media-center"
version = "0.1.0"
description = "A desktop media center application for Linux with streaming-style interface"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-3.0"}
authors = [
    {name = "Videoteka Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
    "mutagen>=1.47.0",
    "PyMuPDF>=1.23.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']


